+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+* 0.0.3 (2014-09-16)
+
+ * updated document storage creation url
+ * added revert to default document storage method
+
 * 0.0.2 (2014-09-05)
 
  * responses returns a json